src/requirements.txt
fastapi==0.68.0
sqlalchemy==1.4.42
databases==0.8.0
python-dotenv==0.19.2
fastapi_limiter==0.1.5
cachetools==5.3.2
asyncio==3.4.3
asyncpg==0.28.0
psycopg2-binary==2.9.1
uvicorn